+++ src/main/java/graphics/IDrawable.java
@@ -7,5 +7,5 @@ public interface IDrawable {
 	 * Render this object on a 2D graphics canvas
 	 * @param g the graphics object to draw to
 	 */
-	public void draw(Graphics2D g, int shift, int shifty, double scalex, double scaley);
+	public void draw(Graphics2D g, float shift, float shifty, float scalex, float scaley);
 }
